# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sampler Book Library API",
        "description": "This is the documentation of the Sampler Book Library API, with this API you can check-in and out books after you register as an user",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get the authenticated User data.",
                "operationId": "App\\Http\\Controllers\\UserController::me",
                "responses": {
                    "200": {
                        "description": "The user saved data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/api/user/books": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user's borrowed books",
                "description": "Retrieve the list of books atributed to the user.",
                "operationId": "App\\Http\\Controllers\\UserController::userBooks",
                "responses": {
                    "200": {
                        "description": "Book list.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Book"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Book": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Harry Potter and the Philosopher's Stone"
                    },
                    "isbn": {
                        "type": "string",
                        "maxLength": 10,
                        "example": "1408855895"
                    },
                    "published_at": {
                        "type": "date",
                        "example": "2014-01-22"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "CHECKED_OUT",
                            "AVAILABLE"
                        ],
                        "example": "AVAILABLE"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Book"
                }
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "email": {
                        "description": "User unique email address",
                        "type": "string",
                        "format": "email",
                        "readOnly": "true",
                        "example": "john@foobar.com"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "John"
                    },
                    "date_of_birth": {
                        "type": "date",
                        "example": "1991-01-22"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "access_token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}